% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{bayesqm-defunct}
\alias{bayesqm-defunct}
\alias{run_bayes}
\alias{plot_hyper}
\alias{select_k_peak}
\alias{select_k_sivula}
\alias{compute_dominant_prob}
\alias{compute_dominant_sign}
\alias{compute_threshold_prob}
\alias{classify_membership}
\alias{compute_divergence}
\alias{critical_delta}
\alias{suggest_delta}
\alias{compute_posterior_scalars}
\alias{demo_run}
\alias{plot_elpd}
\alias{make_elpd_diff}
\alias{make_ppc_ridge}
\alias{make_dominant_panel}
\title{Defunct 0.1.0 functions}
\usage{
run_bayes(...)

plot_hyper(...)

select_k_peak(...)

select_k_sivula(...)

compute_dominant_prob(...)

compute_dominant_sign(...)

compute_threshold_prob(...)

classify_membership(...)

compute_divergence(...)

critical_delta(...)

suggest_delta(...)

compute_posterior_scalars(...)

demo_run(...)

plot_elpd(...)

make_elpd_diff(...)

make_ppc_ridge(...)

make_dominant_panel(...)
}
\arguments{
\item{...}{Ignored.}
}
\value{
Always signals an error.
}
\description{
These 0.1.0 entry points belonged to the removed Student-t score-scale
model and now signal an error naming their 0.2.0 replacement:
flag probabilities and signs come from \code{compute_flags()},
distinguishing/consensus verdicts from \code{compute_qdc()}, the critical
difference from the posterior itself, factor enumeration from
\code{fit_ladder()} and \code{select_k()}, and scalar summaries from
\code{as_draws_df()}.
}
\keyword{internal}
