% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{fitted.bayesqm_fit}
\alias{fitted.bayesqm_fit}
\title{Posterior-mean reconstruction on the utility scale}
\usage{
\method{fitted}{bayesqm_fit}(object, ...)
}
\arguments{
\item{object}{A \code{bayesqm_fit}.}

\item{...}{Unused.}
}
\value{
A \verb{J x N} matrix of posterior-mean latent utilities
\verb{F Lambda'}.
}
\description{
Posterior-mean reconstruction on the utility scale
}
