% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{sigma.bayesqm_fit}
\alias{sigma.bayesqm_fit}
\title{Posterior-mean loading scales}
\usage{
\method{sigma}{bayesqm_fit}(object, ...)
}
\arguments{
\item{object}{A \code{bayesqm_fit}.}

\item{...}{Unused.}
}
\value{
Named numeric vector of posterior-mean \code{sigma_k}.
}
\description{
Posterior-mean loading scales
}
